requirements.txt
openai
python-dotenv
pdf2image
slack_sdk
slack_bolt
requests
nest_asyncio
aiohttp
notion-client
awscli
boto3
PyPDF2
PyPDF
pdfplumber
langchain
faiss-cpu
langchain-community
tiktoken